// generated/oldmodels.d.ts
// GENERATED DO NOT EDIT
// This source data is located at '..\src\opcodes\oldmodels.jsonc'
// run `npm run filetypes` to rebuild

export type oldmodels = {
	colanimcount: number,
	unkcount0: number,
	uvcount: number,
	indexbufsize: number,
	zsize: number,
	ysize: number,
	xsize: number,
	flag1: boolean,
	hascolors: boolean,
	flag3: boolean,
	hasalpha: boolean,
	bonebyte: number,
	extraflags: number,
	texmapcount: number,
	byte4: number,
	facecount: number,
	vertcount: number,
	header1: number,
	header2: number,
	modelversion: number,
	texmap_vertcount: 0,
	texmap_projectioncount: 0,
	texmap_transsize: 0,
	texflags: ({
			type: 0,
			vertindex: number,
		}|{
			type: 1,
			projection: number,
			speed: number,
		}|{
			type: 2,
			projection: number,
			speed: number,
		}|{
			type: 3,
			projection: number,
			speed: number,
		})[],
	vertflags: Uint8Array,
	face2: Buffer | null,
	tritype: Uint8Array,
	boneids: Uint8Array | null,
	unk1: Buffer | null,
	tridata: Buffer,
	alpha: Buffer | null,
	indexbuffer: Buffer,
	material: Uint16Array,
	uvs: Buffer,
	colors: Uint16Array | null,
	posx: Buffer,
	posy: Buffer,
	posz: Buffer,
	texmap_verts: [
		number,
		number,
		number,
	][],
	texmap_projections: {
		normal: [
			number,
			number,
			number,
		],
		scale: [
			(number|number),
			(number|number),
			(number|number),
		],
		rotation: number,
		direction: number,
	}[],
	texmap_translates: number[],
	particles: {
		texture: number,
		faceid: number,
	}[] | null,
	effectors: {
		effector: number,
		vertex: number,
	}[] | null,
	billboards: {
		unk1: number,
		unk2: number,
		unk3: number,
		unk4: number,
	}[] | null,
	texuvs: {
		headoffset: number,
		indexsize: number,
		datasize: number,
		coordcount: number,
		skipped: Buffer,
		offsets: Buffer,
		uvdata: Buffer,
	} | null,
};